Polishing: replace duplicate code by util static function
Mono<byte[]> collectByteArray(Flux<byte[]> bytes) is pulled out to a static
method in OperationUtils so that it is available to the four places it is used
(twice in integration tests and twice in Reactor unit tests). This actually
increases the runtime jar (util is available in production) but not the
runtime footprint. It is a reasonably useful method, however, and reduces the
interdependencies in the test code.