Fill out Gateway Data
Previously, the Gateway Data type was listed as `null` and deprecated in the
documentation.  This meant that we had to guess at the type.  This guess was
`String` and it turns out this is wrong causing deserialization failures. This
change updates the code to be the full type now that we've determined what the
type should be.