Cleanup Logging
Previously, if some part of cleanup failed, only a stack trace and possibly
the request debug logging would be output.  The values within this output
would be UUIDs which are very hard to use for visual inspection and
troubleshooting.  This change adds error logs to each type of cleanup
attaching human-readable names to each.