Strongly Type gateway_data
The documentation for the gateway_data field in various payloads wasn't
particularly revealing requiring us to use a String to represent it.  A user
has an example of the contents of that field and it ends up being a
Map<String, Object>.  This change updates the API to properly represent that
fact.

[#131015779][resolves #566]