Fix Get Application Issue
Previously if an application had failed to stage due to a buildpack compile error,
and a Get Application call was made, we threw a CF-BuildpackCompileFailed exception.
The CLI, which we use as a model, ignores the error and reports correct details of
the application but with no instances running. This commit replicates this behaviour.

[#132208379][fixes #573]