Updates for newer 1.7 builds
This change makes a couple of updates for newer versions of PCF 1.7.
Specifically, this disables the enforcement of failing on unknown properties.
Since the 2.0.x line won't add new fields to match the breaking CAPI changes,
it would fail forever.  This change disables that enforcement (for 2.0.x only)
since it's no longer applicable.