Update Token Request
This change updates the AbstractUaaTokenProvider to be more explicit about
what the request accepts as a response.  This is due to the fact that the one-
time passcode behavior on the UAA doesn't accept the more liberal header.
This change ensures that the client works with a wider variety of UAA
versions.