Release Script Polishing
Previously, the release script assumed that it was run from the cf-java-client folder.
This does not match with the expectation that it will be run as part of a Concourse
pipeline.  This change updates the code so that the working directory is in the expected
location, as well as updating all calls to Maven to use the one packaged within the
projects themselves.