Add Tests for UAA Authorizations
The current state of uaa authorizations has changed significantly since it was first
implemented in cf-java-client. This commit adds tests for missing functionality, and
references to the tasks that should implement each piece of functionality.

It also renames some parts of our current implementation in preparation for the planned
changes.

[#120471689]