Add Conditional Test Support
Previously, all tests ran regardless of the version of CF that the test suite
was running against. We've recently come up against a situation where there
are tests that will never pass on older versions of CF.  This change updates
the project to include the ability to filter tests based on the version of CF
being tested against.