Allow Nulls in Map Values
Previously, Maps in request and response types, could be `null`, but could
never contain keys with null values.  This was due to restrictions in
Immutables.  In the latest release, a new feature arrived that would allow us
to use an annotation to indicate that a Map could contain keys with null
values.  This change updates to the latest version of Immutables, and updates
all the code to deal with this new version.  In addition, it updates all of
the Maps, with the annotation so they should now allow nulls.

[resolves #574]