Token Parameters as Payload
Previously, when a token was negotiated, it used query parameters to pass
along the request payload.  While this was permitted by spec, it did mean that
the debug logging easily showed sensitive information on a regular basis. This
change updates the design of the included TokenProviders to pass the same
information as part of the payload (also permitted by spec).

[resolves #571]