Manual Error Handling
Previously, reactor-netty returned an error for any client failure status code
(i.e. 4XX or 5XX) that contained a reference to the response that came along
with that error code.  This recently changed so that that response was no
longer included.  Instead, the expected behavior is to disable the automatic
mapping to exceptions and instead manually inspect the response, turning it
into an error when the status code is a client failure.  This change updates
our code to do this.