Client V3 Error Type
Previously, the Reactor implementation of the API assumed that the error type
for V3 APIs was the same as V2 APIs.  While this was true in the past, it no
longer is.  This change creates a V3-specific type and ensures that incoming
errors are mapped to that type for all V3 APIs.

[resolves #605]