Add Client Get Security Group Functionality
This commit adds the get security group function in the client.

It also adds all missing client security group integration tests, renames some methods
and fields for consistency with the rest of the application, and adds a new field in the
base security group object that was part of a recent CF update.

[resolves #607]