Enlarge Connection Pool
Previously, the connection pool was the same size as the number of threads on
a system.  This change updates that value to double the number of threads on a
system so that there is some potential for single-thread connection
concurrency in all systems, especially in systems with only a single thread.

[resolves #606]