Close Connection on Cancel
The Router event stream is an odd kind of connection that needs to be
completely closed by the client when consumption is finished.  This change
updates the operation to explicitly close the connection when canceled.  It
doesn't handle the Future part of closing, so this is almost certainly
insufficient for the long term.