User-Agent
Previously, outbound requests did not identify themselves with a User-Agent.
This made troubleshooting requests from the server side more difficult than it
needed to be.  This change adds a User-Agent of `Cloud Foundry Java
Client/<VERSION>` to all requests.  In the case that the version cannot be
determined, `unknown` is used.

[resolves #613]