Routing Root URI
Previously the root of the routing API was assumed to be the same as the CF
root and all URIs contained /routing at the start.  This was incorrect as the
root was actually CF root plus /routing.  Therefore if the routing API was
ever moved, it would break the client.  This change roots the client at the
proper URI.