Disable Connection Pooling
Currently, connection pooling in reactor-netty is enabled by default during
use of the client.  We suspect that there may be some issues with this
connection pooling and as part of testing, we're doing to disable it
universally.  If this change improves the behavior of the client, then it will
stay this way, for all users; i.e. it will not be a configurable option.  At
the time that we have more confidence in the connection pooling, it will be
re-enabled for all users, once again without a configuration property.