Remove BOSH Support
After some further thought, we're removing this BOSH support.  The
documentation of the API, while of high quality is clearly incomplete.  That
coupled with public statements from the BOSH lead that the API isn't intended
to be a widely used entry point means that we shouldn't add support at this
time.  If the API does become a first-class citizen, we'll add it back in.