Resource Matching
Previously, when the push operation was used to send applications to the
server, every single byte was sent, ever time.  The Cloud Foundry server
supports a mechanism for determining if parts of your artifact have already
been uploaded, and if so, you can skip uploading them.  This change takes
advantage of that resource matching API and comes into parity with the Cloud
Foundry CLI.

[resolves #677]