Lazy and Shared Wordlists
Previously, the wordlists used for pushing applications with random names was
loaded eagerly and individually for each instance of them that were created.
Since they were only rarely used anyway, this change ensures that they are
lazily initialized and when they are, they are shared between all instances
that might use them.  This should lower the amount of memory used in cases
where the wordlists aren't used (most cases) and in cases where more than a
single instance is created.

[resolves #676]