Move expensive operations into elastic threads
Previously, all operations took place on the same thread that subscription
happened on.  While this still a reasonable approach for most invocations,
there are a small number of behaviors (e.g. hashing file contents, creating a
filtered archive) that are best moved off the subscribing thread.  Since the
subscribing thread is often the network thread from a previous request (i.e.
chained operations), this has the effect of releasing the network threads
quickly and allowing them to process further requests and responses.