Protect Against Empty Strings in Organization and Space
Previously, organization and space values in the operations layer could be
valid or null.  Since not all operations required an organization or space,
the ids for these artifacts would be lazily resolved, potentionally to an
error if an operation needed the value, but it wasn't configured.  However, an
empty String (meaning, no characters or only whitespace characters), would
count as a "legal" organization or space name, potentially leading to
unintended id resolutions instead of error messages.  This change updates the
code to ensure that those values either have some non-whitespace text or are
considered errors when used, expanding on the null-only check.

[resolves #678]