Improve Exception Stacks
Previously, when exceptions were created within the client, they didn't have
much in the way of useful stack traces to indicate where the exception
originated.  This was partly due to the asynchronous nature of the client, but
also due to the fact that Reactor did not expose that information in a way
that use useful to clients.  With the addition of .checkpoint() to reactor-
core, we now have a way to include both an indication of where in our code the
exception originated, but also where in the users code created the flow which
eventually resulted in the exception.  This should help users of the client
troubleshoot issues in their applications.