Remove NIO File Handling
After struggling mightly with a memory leak around byte[]s and ByteBuffers,
this change removes all use of NIO file handling.  Specifically it appears
that there are some significant memory leaks in ZipFileSystem that cannot be
overcome without removing it's usage.  This change does that resulting in a
significantly better memory profile.