Negotiate Token on First Request
Previously, tokens would not be negotiated until the first 401 was
experienced.  In general this meant that if you were calling an API that
didn't require a token (e.g. /v2/info), no token was ever negotiated. However,
it turned out that there were certain APIs (e.g. /oauth/authorize for SSH
codes), that when called without a token, did not return a 401, but rather
behaved in a way that was correct, but for a different request/response
interaction.  Since this did not work and the number of authenticated APIs
vastly outweighs the number of unauthenticated APIs (~500/~3), we'll negotiate
a token on the first request no matter what.

[resolves #682]