Enlarge Maximum JSON Payload Size
Previously, the newly working JsonObjectDecoder was instantiated using its
default maximum payload size of 1MB.  Certain CF payloads, specifically out of
the UAA can exceed this size.  This change ups that maximum limit to 100M
which should cover all current payloads.  This may need to be enlarged in the
future.