Connect Timeout
Previously, the DefaultConnectionContext allowed the configuration of a socket
timeout.  This didn't make any sense since the socket timeout is a server-side
configuration.  This was the result of a bad conversion from
ClientOptions.timeout(int) when that API was removed.  This change updates the
name of the configuration parameter to ConnectTimeout and sets the appropriate
configuration to match.