Service Instances Integration Tests
This change adds the create and elete integration tests for the service
instances.  These were done because of a bug report that service instance
deletion with accepts_incomplete did not work the way we thought it did.  Lo
and behold, this was true and required extensive rework of service instance
deletion was required as there are now multiple, completely disjoint payloads
that can be returned from service instance deletion.  In addition a new
utility class was added, so that users can wait for last operations to
complete before proceeding; it's not as nice as JobUtils, but then
LastOperations isn't as nice as Jobs.

[resolves #686][#637]