Test Improvements
Previously the user removeManagedOrganization test would fail on PCF 1.7 and
1.8.  This was due to the fact that in those versions no user was associated
as a manager with a new organization.  This change associates the current user
as a manager for the "default" organization after it is created.

In addition to this, the ServiceBrokerTests have been updated to service
brokers after they have been created to account for resource constrained test
environments like pcfdev.

[resolves #695]