Get Refresh Token
A previous change removed the refresh token retrieval API that was used by the
STS team.  This API needed to be reinstated, but we took the opportunity to
improve it a bit.  This change makes that method return a `Flux` allowing a
subscriber to be notified whenever the value changes rather than having to
poll for changes.

[resolves #700]