Caching Refresh Token for Listeners
Previously, the refresh token caching mechanism for listeners used a
DirectProcessor.cache(1).  The had the effect of only caching a value if there
was already a subscriber attached.  As this should return a value on
subscription, regardless of the last time there was a subscriber, this change
updates the code to use a ReplayProcessor which guarentees the "last" value is
always available.