Use Is Filters for Collections of Size One
Previously, all V2 filters on collections used the "In" filter style (key IN
value1,value2).  However, there is a fundamental flaw in that you cannot
safely query for a value with a comma in it because there's no way to escape
the comma within a single value.  This deficiency is not exhibited by the CLI,
because it always uses the "Is" filter style (key:value) and never queries for
more than a single argument.  In practice this is true for nearly every query
done against the cloud controller, so adopting the same pattern is reasonable
for the client.  This gives the flexibility of querying for a single value
with a comma or multiple values with commas, but cannot allow you to query for
multiple values with commas; there is no fix for this coming on the server and
cannot be compensated for on the client.

[resolves #714]