Refresh Token when Getting Username
Previously, when the username was extracted from the current token, if it had
expired at some point that was propagated as a failure.  This change updates
the code such that if there is token expiration failure, the token is
refreshed and the username is extracted again.  If this does not work, an
error is propagated.

While working on the UsernameProvider, I also updated how signing keys were
acquired.  The new implementation deals with signing key rotation by
attempting to use a cached copy of the keys where possible, but when a key is
requested that does not exist, it refreshes the collection of signing keys and
returns the newly matching one.  If this refresh doesn't find the specified
signing key, an error is propagated.

[resolves #718]