Improve Response Logging
This change improves the way that response logging is performed.  It removes
the use of elapsed() which times until the item is passed along, rather than
when the Publisher completes.  It also removes the use of Spring for something
as simple as comma-delimiting a collection.  Once https://github.com/reactor
/reactor-netty/issues/65 is fixed, this will automatically return more
accurate values.