Add Example Configuration
This change adds an ExampleConfiguration that can be used for quickly building
example code.  This will typically be used when reproducing issues or
demonstrating functionality.  All standard beans are exposed, but lazily
initialized, so if your example doesn't require them, you won't be required to
configure them.  The "example" logger (and it's children) are configured to
DEBUG.

Configuration can be achieved by setting the following properties when
required:

* example.apiHost
* example.username
* example.password
* example.organization
* example.space

The ExampleConfiguration should typically be used like this:

@Component
final class Example implements ApplicationRunner {

    private final Logger logger = LoggerFactory.getLogger("example");

    public static void main(String[] args) {
        new SpringApplicationBuilder(Example.class, ExampleConfiguration.class)
            .run(args);
    }

    @Override
    public void run(ApplicationArguments args) throws Exception {

    }

}