Add Index to InstanceDetails
Previously, the index of an instance in an application's instance details was
inferred from its position in the list.  While this was correct, it was not
the most obvious way to communicate this information.  This change adds an
index field to the InstanceDetails type in order to make the index explicit.

[resolves #724]