Default Permissions Used on Non-POSIX Fileystems
Previously, there was an expectation that POSIX file permissions could be
returned from all file systems.  In the case where the file system did not
support this (e.g. ZipFileSystem), null would be returned.  It turns out that
the WindowsFileSystem actually throws and UnsupportedOperationException when
asked for POSIX file permissions (not an unreasonable behavior).  This change
updates the code to ensure that the file system supports POSIX file
permissions before asking for them and if it doesn't, returns the default file
permissions.

[#723]