Protect NetworkLogging Against Failure
Previously, the NetworkLogging class assumed that if you got to time for
printing a response, the response had been generated.  It turns out that there
is no guarantee that a response is generated and the code needs to protect
against this, lest an NPE be generated.  This change adds this protection.