Operations Logging
Previous diagnosis of operations indicated that there just wasn't enough
context about what operations were running in relation to all of the client-
level network requests.  This change is the first, tactical, solution to that
problem.  It adds an operations logger that logs start and finish of
operations and will indicate boundaries on network requests that belong to
operations.