Improve Support for Updating Application Routes
Previously if an application was pushed with a route, and then pushed a second time with
no route, we would see the lack of a route definition in the manifest and create a default
route based on the app name to associate it with. This is incorrect - if an app already
has a route it does not need a default route. This commit brings the client into line with
the CLI in this regard.