ConnectionContext Cleanup
This change adds a dispose() method on the DefaultConnectionContext in order
to cleanup any cached resources like connection and thread pools. This method
is also marked as @PreDestroy so any DefaultConnectionContext created as a
Spring (or EJB) bean will be automatically disposed as part of the standard
application lifecycle.  This implementation is a temporary solution until
https://github.com/reactor/reactor-netty/issues/88 is implemented and
available.

[resolves #745]