Relative Paths to Relative Manifest Paths
Previously, if a relative path to a manifest was used and that manifest
contained a relative path to an artifact an NPE was thrown.  This was because
getting the parent of a top-level relative path returns null in the Path API.
This change immediately resolves any relative paths to an absolute path
allowing further resolutions to work properly.

[resolves #766]