Add Tests For Buildpacks
This commit adds integration and other tests for buildpacks at the client and operations
levels. It also adds the ability to upload a buildpack from a directory (rather than a
zip file) at the client level, for use at the operations level. Finally it removes an
unnecessary field from the create buildpack operation.

[resolves #783]