Inconsistent Mutex
A user reported that occasionally, an NPE would be returned during the sorting
of log messages when streaming their output.  It would have appeared based on
the code that no nulls could exist within the system and therefore there
should be no opportunity for a null to cause an NPE.  However, looking over
the code, it became apparent that the mutex used for addition was different
than the mutex used for compare and remove.  This change ensures that all
access to the queue of elements uses the same mutex.