Work Around Reactor Netty Bug
Currently, there is a bug in Reactor Netty [reactor/reactor-netty#113] where
disabling error mapping for websockets does not work.  The Reactor team has
chosen to only fix this in an API-incompatible release that will not arrive
for months.  In an effort to ensure that CF Java Client customers are not
impacted by this a workaround has been added that intercepts the incorrectly
generated error.

[resolves #774]