Manifest Reading NPE
Previously most of the ApplicationManifestUtils code defensively requested
keys from the manifest payload.  However, there were two places where it
expected a key to exist and would NullPointerException if they didn't.  This
code changes those two locations to be defensive as well, throwing
IllegalStateExceptions when the manifest does not contain the required keys.

[resolves #790]