Introduce a RootProvider
Previously, when attempting to connect to a standalone server (e.g. a
standalone UAA), you were still required to have a Cloud Controller in order
to find out where to negotiate OAuth tokens. This really wouldn't work because
it was more than likely that you didn't actually have a Cloud Controller to
connect to and even if you did, it probably wasn't managing the tokens for the
UAA you were trying to connect to.

In order to solve this issue a new abstraction, the RootProvider, has been
created. This type encapsulates the lookup of URI roots in the system. In the
existing case, there is a InfoPayloadRootProvider that looks up values in the
/v2/info payload of a Cloud Controller. In this standalone server case, you'd
configure a SingleEndpointRootProvider. This RootProvider always returns the
same URI root, regardless of what is asked for.

[#791]