Handle 170001 StagingError from CC when fetching application instance
The current implementation of getApplicationInstances handles errors like
CF_BUILDPACK_COMPILED_FAILED and CF_INSTANCES_ERROR that can be returned from
CC if the staging previously failed. However, the general StagingError (error
code 170001) is not handled and will throw an exception when fetching
application instances.

This change adds support for this error

[#799]