Extract Endpoints from Endpoint Root
Previously all endpoints were extracted from the /v2/info endpoint. With the
forthcoming NetworkingClient, we were looking at a client that was not listed
in that endpoint. This change adds a RootPayloadRootProvider to look to the
new location for this information. For backwards compatibility we've also
introduced a new DelegatingRootProvider that starts by looking at the root
payload and falls back to the info payload if a key is not found.

[#801]