Add unit test for failing manifest inheritance case
There are existing tests for manifest inheritance and common properties, but there is an assumption that the application will be named in the base manifest. This is not documented[1] and is not an assumption held by the CF CLI - the test case added works with the CF CLI. This will need to be fixed to allow the Cloud Foundry Jenkins plugin[2] to easily support this use case.

[1] https://docs.cloudfoundry.org/devguide/deploy-apps/manifest.html#multi-manifests
[2] https://issues.jenkins-ci.org/browse/JENKINS-47271