More Direct Memory Updates
Previously, there was some thought that the problem with direct memory
allocation was related to the size of the chunks that the allocator created.
Experimentation has proven that the issue was totally related to the change
made to Netty 4.1.14.Final that caused it to use an unbounded queue to store
incoming data and not provide feedback to callers that more data was
unwarranted. This change removes a lot of the allocator-related changes from
the code (leaving the monitoring support) and fixes the version of Netty back
to 4.1.13.Final until a later release with a fix to this problem is released.
This change also makes use of the recently backported sendFile API in
reactor-netty.