Manifest Common Inheritance
Previously, the ApplicationManifestUtils did not properly handle inheritance
when the base manifest did not contain an applications section.  This behavior
was supported by the CLI, so the client needed to change to match that.  In
order to support this behavior, the basic inheritance engine was updated such
that if now resolves a canonical YAML manifest first, and then creates types
based on that, rather than creating multiple application manifests and
resolving them in a chain.

[resolves #821]