Log Request and Response Payloads
Previously it was pointed out that while having the request and response URIs
and status codes logged as part of debugging, sometimes more information was
needed.  Specifically having the request and response payloads was useful.
This change adds a TRACE level logging to the request and response loggers
that contains the toString() representation of any request and response
objects.

[resolves #827]