Handle Empty Root Payloads
A recent change updated how root URIs were found to prefer the payload of the
root endpoint (/) to  the payload of the info endpoint (/v2/info).  However,
there are some older CF installations (such as Stackato) that not only don't
have that root endpoint, but the URI is redirected to some other location.
This change updates the RootProvider implementations to properly handle this
"empty payload" case and fall back properly.

[resolves #817]