Less SSL Trusting
Previousy, when trusting certificates, if the certificate was already trusted,
but under a different host and port (e.g. login.local.pcfdev.io was previously
trusted as api.local.pcfdev.io) then the host and port would never be added to
the list of already trusted hosts and ports.  This meant that every time that
host and port needed to be trusted again, the check would happen again rather
than short circuiting.  This change ensures that this host and port are added
as long as the certificate presented has been previously or is now trusted.