Periodically Invalidate Cached Items
Previously, if there was an error when a cached item was first calculated,
that error would be cached forever.  This change adds a `cacheDuration()`
method to the DefaultConnectionContext and DefaultCloudFoundryOperations in
order to configure how long _anything_ should be cached.  This could be
something like  how long the root payload is cached or the organization and
space ids.  The default is cached indefinitely, but can be configured.

[#749]