Handle JMX Collisions
Previously, if two DefaultConnectionContexts were created in a single JVM,
registration of the JMX ByteBuf monitors would cause an ugly message to be
logged.  While this wasn't terminal, it wasn't a great experience.  In
addition, the possibility of multiple DefaultConnectionContexts is a valid one
as you might need to connect to multiple host/port endpoints.  This change
updates the name of the MBeans to include endpoint to properly handle the
valid case of multiple DefaultConnectionContexts for multiple endpoints.  It
also warns about multiple duplicate DefaultConnectionContexts for the same
endpoint and removes the previous MBean, replacing it with the new one.  It
also unregisters the MBean during disposal as part of normal cleanup.

[resolves #836]