ByteBuf Scope Improvements
Previously, two parts of the code used asInputStream() to get access to the
contents of a ByteBufFlux. The contents of these types were used outside of
the flatMapMany() they were created in leading to some potential non-closures
of those InputStreams which would leave ByteBufs retained and never released.
This likely wasn't really an issue, but a code review brought them up and the
scoping has now been changed. The InputStreams are now read to exhaustion and
closed in the same operator they were created in, resulting in a guaranteed
closure.