Fix Updated At Issue
Previously we assumed that v3 resources would always have their `updatedAt` field
populated. In general this is true, but there are instances in production where it isn't
(for example, where an organization has existed since much earlier versions). To allow
for this, this commit allows for `updatedAt` to be empty in v3, in line with how we
already handle v2.

[resolves #871]