Improve User Cleanup
Previously the basic test user created during integration test runs was not being cleaned
up during startup/shutdown. This was because it derived its identifying feature (username)
from the corresponding uaa user, and that uaa user was being cleaned up first, removing
the username and leaving us unable to identify the cf user as suitable for deletion. This
commit fixes the ordering problem.

[resolves #885]