Reduce Port Clashes
Previously we used a random number for port selection, which sometimes led to clashes
with existing processes. This commits changes port selection to try to avoid used ports.

Note that the reliability of the selected method, `SocketUtils.findAvailableTcpPort()`,
has been questioned, so this change may be backed out if it does not result in an
improvement.

[resolves #863]