Improve Process Backwards Compatibility
Previously the V3 api assumed that no process would have a health check type of `none`,
which was deprecated in V2. Until the world is V3 only, however, `none` can still exist.
This commit updates the client to allow this value.

It also adds integration tests for Process, and includes updates in the underlying api.

[resolves #909]