Manifest Memory Units
Previously when the client created a manifest from an existing app it would write the disk
and memory values as simple integers, on the assumption that both fields represent
megabytes. This assumption is correct, but is not according to the spec, which requires a
unit indicator. This causes an error if the manifest is then used by the CLI. This commit
makes explicit the assumed unit by appending 'M' to the value of each field.

It also modifies how paths are read from manifests to improve compatibility with Windows
file references. Although it has been tested, feedback on this change is welcome.

[resolves #911]