Handle Optional Internal Field
A previous commit handled the new option for internal domains by filtering out any marked
as internal in some circumstances. This caused a failure when testing against all but the
latest version of CF, as the filter logic caused an NPE if the internal field did not
exist. This change handles the optionality of that field.