Handle Host Names with Dots
In some circumstances an application's name is used to form a host name, for example
when assigning a random route (where the route would be called something like
appname-random-words.example.com). If the application name includes dots (such as
appname-2.0.1) we would pass a failure to the caller on push, as CF rejects host names
with dots. The CLI handles this by stripping out dots, (giving, say, appname-201). This
commit replicates that behaviour in our operations layer.