Time Zone Parsing Error
An internal date conversion utility was incorrectly configured to consider only a single
character of a time zone offset value. This worked fine if the user was in, say,
California (offset -8), but caused a 30 or 45 minute error for places such as much of
India (offset +5:30) or Nepal (+5:45). This commit takes the full offset into account.

[resolves #925]