Fix Direct Memory Leaks in Doppler Client
Previously the versions of Reactor that we used had a behavior where, in the
case of a cancellation of a subscription, any element that was caching data
wouldn't be notified and they'd be at risk of creating a memory leak (think
buffers and streaming codecs).

This change makes use of new Reactor features that allow notification of
cancellation and cleanup of state).

[resolves #1019]

Signed-off-by: Ben Hale <bhale@pivotal.io>