Resubscribe on Token Expiration
Before the recent rebase onto modern versions of Reactor Netty ([#984]), when
the Client received a 401 Unauthorized, it would invalidated any existing
token and then resubscribe with the same request causing a new token to be
negotiated the request to be retried, transparently to the user.

During the rebase, this functionality was lost (an easy mistake to make given
that there are no tests for it...) resulting in invalidation of the token, but
no re-subscription and therefore the failure was exposed to the user.

This change re-introduced the previous re-subscribe behavior.

Signed-off-by: Ben Hale <bhale@pivotal.io>