Resubscribe on Token Expiration II
Well, it turns out that my first attempt at this was only partially correct.
The code did re-subscribe appropriately on an invalid token, but the
subscription ended up being lost within a callback that returned a
Mono.just();  This change forces the subscription chain out of that callback
and to the request/response pair itself, resulting in a proper transparent
re-request.

A minor change was also made to how the trigger for resubscribe happens.

Signed-off-by: Ben Hale <bhale@pivotal.io>