Selective JSON Decoding
Signed-off-by: Ben Hale <bhale@pivotal.io>

Improve Error Handling on Startup

Previously we assumed that any value returned from CF would be formatted as json. This
isn't true if we have certain error conditions, for example if a 404 is returned when we
try to access the /v2/info endpoint. These scenarios produced unhelpful json-based errors
that hid the actual problem.

With this commit, while we will still produce an error (because we need the json response
to continue), the error should be informative.

Signed-off-by: Paul Harris <harrisp@vmware.com>